test: split large inline test modules
Move bulky inline Rust test blocks into sibling *_test.rs files so the implementation modules stay easier to scan without changing test coverage or production visibility.
